requirements.txt
# Web UI
streamlit>=1.28.0

# TTS - ElevenLabs
elevenlabs>=1.0.0

# AWS
boto3>=1.28.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
